MDEV-34724: Skipping a row operation from a trigger
Implementation of this task adds abililty to raise the signal with
SQLSTATE '02TRG' from a BEFORE INSERT/UPDATE/DELETE trigger and handles
this signal as an indicator meaning 'to throw away the current row'
on processing the INSERT/UPDATE/DELETE statement. The signal with
SQLSTATE '02TRG' has special meaning only in case it is raised inside
BEFORE triggers, for AFTER trigger's this value of SQLSTATE isn't treated
in any special way. In according with SQL standard, the SQLSTATE class '02'
means NO DATA and sql_errno for this class is set to value
ER_SIGNAL_NOT_FOUND by current implementation of MariaDB server.
Implementation of this task assigns the value ER_SIGNAL_SKIP_ROW_FROM_TRIGGER
to sql_errno in Diagnostics_area in case the signal is raised from a trigger
and SQLSTATE has value '02TRG'.

To catch signal with SQLTSATE '02TRG' and handle it in special way, the methods
 Table_triggers_list::process_triggers
 select_insert::store_values
 select_create::store_values
 Rows_log_event::process_triggers
and the overloaded function
 fill_record_n_invoke_before_triggers
were extended with extra out parameter for returning the flag whether
to skip the current values being processed by INSERT/UPDATE/DELETE
statement. This extra parameter is passed as nullptr in case of AFTER trigger
and BEFORE trigger this parameter points to a variable to store a marker
whether to skip the current record or store it by calling write_record().

+++ sql/sql_delete.cc
@@ -276,7 +276,7 @@ int update_portion_of_time(THD *thd, TABLE *table,
 
   if (likely(!res) && table->triggers)
     res= table->triggers->process_triggers(thd, TRG_EVENT_INSERT,
-                                           TRG_ACTION_AFTER, true);
+                                           TRG_ACTION_AFTER, true, nullptr);
   restore_record(table, record[1]);
   if (res)
     table->file->restore_auto_increment(prev_insert_id);
@@ -874,15 +874,22 @@ bool Sql_cmd_delete::delete_from_single_table(THD *thd)
     {
       void *save_bulk_param= thd->bulk_param;
       thd->bulk_param= nullptr;
+
+      bool trg_skip_row= false;
+
       if (!delete_history && table->triggers &&
           table->triggers->process_triggers(thd, TRG_EVENT_DELETE,
-                                            TRG_ACTION_BEFORE, FALSE))
+                                            TRG_ACTION_BEFORE, false,
+                                            &trg_skip_row))
       {
         error= 1;
         thd->bulk_param= save_bulk_param;
         break;
       }
       thd->bulk_param= save_bulk_param;
+
+      if (trg_skip_row)
+        continue;
 
       // no LIMIT / OFFSET
       if (returning && result->send_data(returning->item_list) < 0)
@@ -916,7 +923,8 @@ bool Sql_cmd_delete::delete_from_single_table(THD *thd)
 	thd->bulk_param= nullptr;
         if (!delete_history && table->triggers &&
             table->triggers->process_triggers(thd, TRG_EVENT_DELETE,
-                                              TRG_ACTION_AFTER, FALSE))
+                                              TRG_ACTION_AFTER, false,
+                                              nullptr))
         {
           error= 1;
           thd->bulk_param= save_bulk_param;
@@ -1263,12 +1271,19 @@ int multi_delete::send_data(List<Item> &values)
 
     if (secure_counter < 0)
     {
+      bool trg_skip_row= false;
+
       /* We are scanning the current table */
       DBUG_ASSERT(del_table == table_being_deleted);
       if (table->triggers &&
           table->triggers->process_triggers(thd, TRG_EVENT_DELETE,
-                                            TRG_ACTION_BEFORE, FALSE))
+                                            TRG_ACTION_BEFORE, false,
+                                            &trg_skip_row))
         DBUG_RETURN(1);
+
+      if (trg_skip_row)
+        continue;
+
       table->status|= STATUS_DELETED;
 
       error= table->delete_row();
@@ -1279,7 +1294,8 @@ int multi_delete::send_data(List<Item> &values)
           thd->transaction->stmt.modified_non_trans_table= TRUE;
         if (table->triggers &&
             table->triggers->process_triggers(thd, TRG_EVENT_DELETE,
-                                              TRG_ACTION_AFTER, FALSE))
+                                              TRG_ACTION_AFTER, false,
+                                              nullptr))
           DBUG_RETURN(1);
       }
       else if (!ignore)
@@ -1442,13 +1458,19 @@ int multi_delete::do_table_deletes(TABLE *table, SORT_INFO *sort_info,
   bool will_batch= !table->file->start_bulk_delete();
   while (likely(!(local_error= info.read_record())) && likely(!thd->killed))
   {
+    bool trg_skip_row= false;
+
     if (table->triggers &&
         unlikely(table->triggers->process_triggers(thd, TRG_EVENT_DELETE,
-                                                   TRG_ACTION_BEFORE, FALSE)))
+                                                   TRG_ACTION_BEFORE, false,
+                                                   &trg_skip_row)))
     {
       local_error= 1;
       break;
     }
+
+    if (trg_skip_row)
+      continue;
 
     local_error= table->delete_row();
     if (unlikely(local_error) && !ignore)
@@ -1467,7 +1489,8 @@ int multi_delete::do_table_deletes(TABLE *table, SORT_INFO *sort_info,
       deleted++;
       if (table->triggers &&
           table->triggers->process_triggers(thd, TRG_EVENT_DELETE,
-                                            TRG_ACTION_AFTER, FALSE))
+                                            TRG_ACTION_AFTER, false,
+                                            nullptr))
       {
         local_error= 1;
         break;
